조편성 삭제 제한 추가

+++ tabs/squad_settings.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from datetime import datetime, timedelta
 
 from PyQt5.QtCore import Qt, QDate
 from PyQt5.QtWidgets import QVBoxLayout, QTableWidgetItem, QTableWidget, QWidget, QButtonGroup, QRadioButton, \
@@ -33,8 +33,8 @@ class SquadSettingsTab(QWidget):
 
         # 사용자 명단 테이블 추가
         self.squad_table = QTableWidget()
-        self.squad_table.setColumnCount(3)
-        self.squad_table.setHorizontalHeaderLabels(["선택", "date_label", "leaders"])
+        self.squad_table.setColumnCount(4)
+        self.squad_table.setHorizontalHeaderLabels(["선택", "date_label", "leaders", "date_text"])
         self.squad_table.setEditTriggers(QTableWidget.NoEditTriggers)  # 기본적으로 편집 불가능
         self.main_layout.addWidget(self.squad_table)
         self.squad_table.verticalHeader().setVisible(False)
@@ -43,6 +43,8 @@ class SquadSettingsTab(QWidget):
         self.load_user_data()
 
     def load_user_data(self):
+        self.delete_button.setEnabled(True)
+
         user_data = self.db.select_team_history()
         self.squad_table.setRowCount(len(user_data))
         self.squad_table.cellClicked.connect(self.on_cell_clicked)
@@ -50,7 +52,7 @@ class SquadSettingsTab(QWidget):
 
         self.radio_button_group = QButtonGroup(self.squad_table)
         self.radio_button_group.setExclusive(True)
-        for row_index, (squad_id, date_label, leader_ids) in enumerate(user_data):
+        for row_index, (squad_id, date_label, leader_ids, date_text) in enumerate(user_data):
             radio_button_cell = self.create_radio_button_cell(self.radio_button_group, row_index)
             self.squad_table.setCellWidget(row_index, 0, radio_button_cell)
             self.squad_table.setColumnWidth(0, 100)
@@ -63,8 +65,11 @@ class SquadSettingsTab(QWidget):
             self.squad_table.setItem(row_index, 2, last_date_item)
             self.squad_table.setColumnWidth(2, 200)
 
+            date_text_item = QTableWidgetItem(str(date_text))
+            self.squad_table.setItem(row_index, 3, date_text_item)
+            self.squad_table.setColumnWidth(3, 200)
+
     def on_cell_clicked(self, row, column):
-        # 첫 번째 열(선택 칸)인 경우 라디오 버튼 선택
         if column == 0:
             radio_button = self.user_table.cellWidget(row, column)
             if radio_button:
@@ -84,11 +89,31 @@ class SquadSettingsTab(QWidget):
         self.load_user_data()
 
     def delete_squad_data(self):
+        current_date = datetime.now()
+        year, month = map(int, current_date.strftime("%Y %m").split(" "))
+        first_of_month = QDate(year, month, 1)
+
+        first_monday = first_of_month
+        while first_monday.dayOfWeek() != 1:
+            first_monday = first_monday.addDays(1)
+
+        if first_monday > current_date:
+            last_date = QDate(year, month, 1).addDays(-1)
+            last_of_month = QDate(last_date.year(), last_date.month(), 1)
+
+            first_monday = last_of_month
+            while first_monday.dayOfWeek() != 1:
+                first_monday = first_monday.addDays(1)
+        this_week_date = (current_date - timedelta(days=current_date.weekday())).strftime("%Y-%m-%d")
+
         selected_row = self.get_selected_row()
         if selected_row is not None:
-            date_label = self.get_date_label(selected_row)
-            self.db.delete_team_history(date_label)
-            self.reload_users()
+            self.delete_button.setEnabled(False)
+
+            date_label, date_text = self.get_data(selected_row)
+            if this_week_date <= date_text:
+                self.db.delete_team_history(date_label)
+                self.reload_users()
 
     def get_selected_row(self):
         for row in range(self.squad_table.rowCount()):
@@ -97,6 +122,7 @@ class SquadSettingsTab(QWidget):
                 return row
         return None
 
-    def get_date_label(self, row):
+    def get_data(self, row):
         id_item = self.squad_table.item(row, 1)
-        return id_item.text() if id_item else None
+        date_text = self.squad_table.item(row, 3)
+        return (id_item.text(), date_text.text()) if id_item else (None, None)
